requirements.txt
amulet-core
Pillow
dotenv